Starting over on comments-data with fresh projects
My previous attempt had created a model 7000068 project and then opened
it with 7000069 and 7000070. This resulted in three repos that were all
connected to each other (started with the same root commit), and so when
Chorus does a Send/Receive to the USB stick with three repos on it, it
was random which project would be selected, which made TestUtil's wizard
mode unusable. This time we're starting with three unconnected projects
(first commits are different) so Chorus won't get confused.